before render fix

+++ src/setup.h
@@ -46,7 +46,7 @@ typedef struct
 
 typedef struct
 {
-	uint8_t material;	// 8 bits
+	uint32_t material;	// 32 bits
 	float friction;		// 32 bits
 	bool isFreeFalling; // 8 bits ?
 	float velocityX;	// 32 bits
@@ -54,14 +54,14 @@ typedef struct
 	int32_t mass;		// 32 bits
 	float spreadFactor; // 32 bits
 	Color color;		// 32 bits
-} Cell;					// 208 bits
+} Cell;					// 232 bits
 
 // Function prototype
 int32_t setup_stuff(int32_t sc_wi, int32_t sc_he, const char *WindowTitle, int32_t log_lvl, bool fullscreen);
 int32_t set_monitor_and_fps(int32_t monitor);
 void sand(Cell (*grid)[ROWS], Cell (*grid_duplicate)[ROWS]);
 void updateWater(Cell (*grid)[ROWS], Cell (*grid_duplicate)[ROWS]);
-void spawnSandBrush(Cell (*grid)[ROWS], int32_t mouseX, int32_t mouseY, int32_t brushSize, int32_t material, bool brushMode);
-Color rand_color_mat(int32_t material);
+void spawnSandBrush(Cell (*grid)[ROWS], int32_t mouseX, int32_t mouseY, int32_t brushSize, uint32_t material, bool brushMode);
+Color rand_color_mat(uint32_t material);
 
 #endif /* SETUP_H */
